requirements.txt
pyttsx3==2.90
SpeechRecognition==3.10.4
urllib3==2.2.2
requests==2.32.3
pygame==2.6.0
google-generativeai==0.7.1
emoji==2.12.1
python-dotenv==1.0.1
elevenlabs==1.3.1